+++ pack.py
@@ -27,7 +27,7 @@ import sys
 import zipfile
 from pathlib import Path
 
-VERSION = "1.0.2"
+VERSION = "1.0.3"
 ROOT = Path(__file__).parent.resolve()